event.display() instead of notification
